Removes downloading mods reference
First of all Thank You for this great guide! I am suggesting this edit because it was confusing reading this guide to see "next step is installing mods" and then "now we are going to install another thing that you can't install with the mod manager. I might be totally confused (I haven't played morrowind with mods before) but just trying to help.